feat(publisher): rework and refactor
- move related code to the pkg
- implement publisher interface
- perform all actions with locking
