+++ path_publish.go
@@ -14,7 +14,6 @@ import (
 	"github.com/werf/vault-plugin-secrets-trdl/pkg/config"
 	trdlGit "github.com/werf/vault-plugin-secrets-trdl/pkg/git"
 	"github.com/werf/vault-plugin-secrets-trdl/pkg/pgp"
-	"github.com/werf/vault-plugin-secrets-trdl/pkg/publisher"
 	"github.com/werf/vault-plugin-secrets-trdl/pkg/tasks_manager"
 	"github.com/werf/vault-plugin-secrets-trdl/pkg/util"
 )
@@ -94,7 +93,7 @@ func (b *backend) pathPublish(ctx context.Context, req *logical.Request, fields 
 		}
 	}
 
-	publisherRepository, err := GetPublisherRepository(ctx, cfg, req.Storage)
+	publisherRepository, err := b.Publisher.GetRepository(ctx, req.Storage, cfg.RepositoryOptions())
 	if err != nil {
 		return nil, fmt.Errorf("error getting publisher repository: %s", err)
 	}
@@ -151,7 +150,7 @@ func (b *backend) pathPublish(ctx context.Context, req *logical.Request, fields 
 		logboek.Context(ctx).Default().LogF("Got trdl channels config:\n%s\n---\n", cfgDump)
 		hclog.L().Debug(fmt.Sprintf("Got trdl channels config:\n%s\n---", cfgDump))
 
-		if err := publisher.PublishChannelsConfig(ctx, publisherRepository, cfg); err != nil {
+		if err := b.Publisher.PublishChannelsConfig(ctx, publisherRepository, cfg); err != nil {
 			return fmt.Errorf("error publishing trdl channels into the repository: %s", err)
 		}
