+++ pkg/publisher/manager.go
@@ -1,0 +1,11 @@
+package publisher
+
+import "sync"
+
+type Publisher struct {
+	mu sync.Mutex
+}
+
+func NewPublisher() Interface {
+	return &Publisher{}
+}
