+++ pkg/publisher/periodic.go
@@ -1,0 +1,14 @@
+package publisher
+
+import (
+	"context"
+
+	"github.com/hashicorp/vault/sdk/logical"
+)
+
+func (m *Publisher) PeriodicTask(_ context.Context, _ *logical.Request) error {
+	m.mu.Lock()
+	defer m.mu.Unlock()
+
+	return nil
+}
